+++ assets/js/index.js
@@ -433,12 +433,7 @@ $(document).ready(function(){
             type : "POST",
             contentType : 'application/json', 
             data : form_data,
-            success : function(result){  
-                // check if user is disabled
-                if(result.accState==0) {
-                    $('#response').html("<div class='alert alert-danger'>Login failed. Your account is currently disabled! Please contact your administrator.</div>");
-                    login_form.find('input').val(''); 
-                } else {
+            success : function(result){                  
                     // check if mfa is enabled
                     if(result.mfa == 1) {
                         // store jwt to cookie
@@ -451,8 +446,7 @@ $(document).ready(function(){
                         // show home page & tell the user it was a successful login
                         showHomePage();
                         $('#response').html("<div class='alert alert-success'>Successful login.</div>");                    
-                    } 
-                }                      
+                    }                       
             },
             error: function(xhr, resp, text){                
                 // on error, tell the user login has failed & empty the input boxes