+++ changelog.md
@@ -94,7 +94,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.0.5] 2021-12-20
 ### Added
 - modified logon api to reflect brute force settings
+- added mail class for sending mails
+- added check for admin notify on disabeling a user due to too many failed logins 
 - brute force feature complete
-